Tests/LibWeb/Text/expected/Crypto/SubtleCrypto-rsa-bad-import-key.txt

RSASSA-PKCS1-v1_5 SHA-1 PRIV - jwk   OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-1 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-1 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-256 PRIV - jwk   OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-256 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-256 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-384 PRIV - jwk   OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-384 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-384 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-512 PRIV - jwk   OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-512 PRIV - pkcs8 OK: DataError: Invalid key
RSASSA-PKCS1-v1_5 SHA-512 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-1 PRIV - jwk   OK: DataError: Invalid key
RSA-PSS SHA-1 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-1 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-256 PRIV - jwk   OK: DataError: Invalid key
RSA-PSS SHA-256 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-256 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-384 PRIV - jwk   OK: DataError: Invalid key
RSA-PSS SHA-384 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-384 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-512 PRIV - jwk   OK: DataError: Invalid key
RSA-PSS SHA-512 PRIV - pkcs8 OK: DataError: Invalid key
RSA-PSS SHA-512 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-1 PRIV - jwk   OK: DataError: Invalid key
RSA-OAEP SHA-1 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-1 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-256 PRIV - jwk   OK: DataError: Invalid key
RSA-OAEP SHA-256 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-256 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-384 PRIV - jwk   OK: DataError: Invalid key
RSA-OAEP SHA-384 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-384 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-512 PRIV - jwk   OK: DataError: Invalid key
RSA-OAEP SHA-512 PRIV - pkcs8 OK: DataError: Invalid key
RSA-OAEP SHA-512 PRIV - pkcs8 OK: DataError: Invalid key
